Code clean up: small refactor to paste code (#2365)
* Refactor paste

* add test
